feat(rust): Add re-registration of OOC SpillTokens
